requirements.txt
Django==3.2.25
psycopg2-binary==2.8.5
djangorestframework==3.15.1
django-environ==0.11.2